requirements.txt
paho-mqtt==1.5.0
psycopg2-binary==2.8.5
